# bcc-orgs/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Orgs API",
        "title": "BCC Orgs API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/api/v1",
    "paths": {
        "/orgs/{orgID}": {
            "get": {
                "description": "Org retrieval is permitted through the use of scopes. For scope definitions go to https://bcc-code.github.io/projects/bcc-membership-docs/data-structures-and-scopes.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orgs"
                ],
                "summary": "Get org by orgID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "orgID",
                        "name": "orgID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Org"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Address": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "countryIso2Code": {
                    "type": "string"
                },
                "countryName": {
                    "type": "string"
                },
                "countryNameNative": {
                    "type": "string"
                },
                "postalCode": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "street1": {
                    "type": "string"
                },
                "street2": {
                    "type": "string"
                }
            }
        },
        "Org": {
            "type": "object",
            "properties": {
                "billingAddress": {
                    "$ref": "#/definitions/Address"
                },
                "legalName": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "orgID": {
                    "type": "integer"
                },
                "postalAddress": {
                    "$ref": "#/definitions/Address"
                },
                "type": {
                    "type": "string"
                },
                "visitingAddress": {
                    "$ref": "#/definitions/Address"
                }
            }
        }
    }
}